+++ shared/constructor/__init__.py
@@ -1,0 +1,1 @@
+"""Metadata constructor: external data processor projects in constructor.db."""